hifi(slice1): fix collectDsSlugs content-key + extract ds-anatomy-map (C2)
Review C2: collectDsSlugs walked screens[].frames but real flow-data carries
nodes under screens[].content, so the anatomy map was always empty. Fix the
key + stale doc comment, and extract collectDsSlugs/buildDsAnatomyMap into
ds-anatomy-map.js so the flow-share assembler can build the map without a
renderer->CLI dependency (assemble-preview re-exports for back-compat).
Adds collectDsSlugs coverage (content traversal, dedupe, nil guards).
